Extract prompts from codex-core
