+++ codex-rs/core/src/client_common.rs
@@ -12,14 +12,6 @@ use std::task::Poll;
 use tokio::sync::mpsc;
 use tokio_util::sync::CancellationToken;
 
-/// Review thread system prompt. Edit `core/src/review_prompt.md` to customize.
-pub const REVIEW_PROMPT: &str = include_str!("../review_prompt.md");
-
-// Centralized templates for review-related user messages
-pub const REVIEW_EXIT_SUCCESS_TMPL: &str = include_str!("../templates/review/exit_success.xml");
-pub const REVIEW_EXIT_INTERRUPTED_TMPL: &str =
-    include_str!("../templates/review/exit_interrupted.xml");
-
 /// API request payload for a single model turn
 #[derive(Debug, Clone)]
 pub struct Prompt {
