+++ codex-rs/core/src/context/realtime_end_instructions.rs
@@ -1,8 +1,7 @@
 use super::ContextualUserFragment;
+use codex_prompts::END_INSTRUCTIONS;
 use codex_protocol::protocol::REALTIME_CONVERSATION_CLOSE_TAG;
 use codex_protocol::protocol::REALTIME_CONVERSATION_OPEN_TAG;
-
-const REALTIME_END_INSTRUCTIONS: &str = include_str!("prompts/realtime/realtime_end.md");
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub(crate) struct RealtimeEndInstructions {
@@ -34,10 +33,6 @@ impl ContextualUserFragment for RealtimeEndInstructions {
     }
 
     fn body(&self) -> String {
-        format!(
-            "\n{}\n\nReason: {}\n",
-            REALTIME_END_INSTRUCTIONS.trim(),
-            self.reason
-        )
+        format!("\n{}\n\nReason: {}\n", END_INSTRUCTIONS.trim(), self.reason)
     }
 }
