+++ codex-rs/core/src/context/realtime_start_instructions.rs
@@ -1,8 +1,7 @@
 use super::ContextualUserFragment;
+use codex_prompts::START_INSTRUCTIONS;
 use codex_protocol::protocol::REALTIME_CONVERSATION_CLOSE_TAG;
 use codex_protocol::protocol::REALTIME_CONVERSATION_OPEN_TAG;
-
-const REALTIME_START_INSTRUCTIONS: &str = include_str!("prompts/realtime/realtime_start.md");
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub(crate) struct RealtimeStartInstructions;
@@ -24,6 +23,6 @@ impl ContextualUserFragment for RealtimeStartInstructions {
     }
 
     fn body(&self) -> String {
-        format!("\n{}\n", REALTIME_START_INSTRUCTIONS.trim())
+        format!("\n{}\n", START_INSTRUCTIONS.trim())
     }
 }
