+++ codex-rs/prompts/BUILD.bazel
@@ -1,0 +1,7 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "prompts",
+    crate_name = "codex_prompts",
+    compile_data = glob(["templates/**"]),
+)
