+++ codex-rs/prompts/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+edition.workspace = true
+license.workspace = true
+name = "codex-prompts"
+version.workspace = true
+
+[lib]
+doctest = false
+name = "codex_prompts"
+path = "src/lib.rs"
+
+[lints]
+workspace = true
+
+[dependencies]
+anyhow = { workspace = true }
+codex-git-utils = { workspace = true }
+codex-execpolicy = { workspace = true }
+codex-protocol = { workspace = true }
+codex-utils-absolute-path = { workspace = true }
+codex-utils-template = { workspace = true }
+
+[dev-dependencies]
+pretty_assertions = { workspace = true }
