+++ codex-rs/prompts/src/agents.rs
@@ -1,0 +1,1 @@
+pub const HIERARCHICAL_AGENTS_MESSAGE: &str = include_str!("../templates/agents/hierarchical.md");
