+++ codex-rs/prompts/src/apply_patch.rs
@@ -1,0 +1,3 @@
+/// Detailed instructions for gpt-4.1 on how to use the `apply_patch` tool.
+pub const APPLY_PATCH_TOOL_INSTRUCTIONS: &str =
+    include_str!("../templates/apply_patch_tool_instructions.md");
