+++ codex-rs/prompts/src/compact.rs
@@ -1,0 +1,2 @@
+pub const SUMMARIZATION_PROMPT: &str = include_str!("../templates/compact/prompt.md");
+pub const SUMMARY_PREFIX: &str = include_str!("../templates/compact/summary_prefix.md");
