+++ codex-rs/prompts/src/lib.rs
@@ -1,0 +1,27 @@
+mod agents;
+mod apply_patch;
+mod compact;
+mod goals;
+mod permissions_instructions;
+mod realtime;
+mod review_exit;
+mod review_request;
+
+pub use agents::HIERARCHICAL_AGENTS_MESSAGE;
+pub use apply_patch::APPLY_PATCH_TOOL_INSTRUCTIONS;
+pub use compact::SUMMARIZATION_PROMPT;
+pub use compact::SUMMARY_PREFIX;
+pub use goals::budget_limit_prompt;
+pub use goals::continuation_prompt;
+pub use goals::objective_updated_prompt;
+pub use permissions_instructions::PermissionsInstructions;
+pub use realtime::BACKEND_PROMPT;
+pub use realtime::END_INSTRUCTIONS;
+pub use realtime::START_INSTRUCTIONS;
+pub use review_exit::render_review_exit_interrupted;
+pub use review_exit::render_review_exit_success;
+pub use review_request::REVIEW_PROMPT;
+pub use review_request::ResolvedReviewRequest;
+pub use review_request::resolve_review_request;
+pub use review_request::review_prompt;
+pub use review_request::user_facing_hint;
