+++ codex-rs/prompts/src/permissions_instructions_tests.rs
