+++ codex-rs/prompts/src/realtime.rs
@@ -1,0 +1,3 @@
+pub const BACKEND_PROMPT: &str = include_str!("../templates/realtime/backend_prompt.md");
+pub const END_INSTRUCTIONS: &str = include_str!("../templates/realtime/realtime_end.md");
+pub const START_INSTRUCTIONS: &str = include_str!("../templates/realtime/realtime_start.md");
