+++ codex-rs/prompts/src/review_request.rs
@@ -5,6 +5,9 @@ use codex_utils_absolute_path::AbsolutePathBuf;
 use codex_utils_template::Template;
 use std::sync::LazyLock;
 
+/// Review thread system prompt.
+pub const REVIEW_PROMPT: &str = include_str!("../templates/review/rubric.md");
+
 #[derive(Clone, Debug, PartialEq)]
 pub struct ResolvedReviewRequest {
     pub target: ReviewTarget,
@@ -130,56 +133,5 @@ impl From<ResolvedReviewRequest> for ReviewRequest {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use pretty_assertions::assert_eq;
-
-    #[test]
-    fn review_prompt_template_renders_base_branch_backup_variant() {
-        assert_eq!(
-            render_review_prompt(&BASE_BRANCH_PROMPT_BACKUP_TEMPLATE, [("branch", "main")]),
-            "Review the code changes against the base branch 'main'. Start by finding the merge diff between the current branch and main's upstream e.g. (`git merge-base HEAD \"$(git rev-parse --abbrev-ref \"main@{upstream}\")\"`), then run `git diff` against that SHA to see what changes we would merge into the main branch. Provide prioritized, actionable findings."
-        );
-    }
-
-    #[test]
-    fn review_prompt_template_renders_base_branch_variant() {
-        assert_eq!(
-            render_review_prompt(
-                &BASE_BRANCH_PROMPT_TEMPLATE,
-                [("base_branch", "main"), ("merge_base_sha", "abc123")]
-            ),
-            "Review the code changes against the base branch 'main'. The merge base commit for this comparison is abc123. Run `git diff abc123` to inspect the changes relative to main. Provide prioritized, actionable findings."
-        );
-    }
-
-    #[test]
-    fn review_prompt_template_renders_commit_variant() {
-        assert_eq!(
-            review_prompt(
-                &ReviewTarget::Commit {
-                    sha: "deadbeef".to_string(),
-                    title: None,
-                },
-                &AbsolutePathBuf::current_dir().expect("cwd"),
-            )
-            .expect("commit prompt should render"),
-            "Review the code changes introduced by commit deadbeef. Provide prioritized, actionable findings."
-        );
-    }
-
-    #[test]
-    fn review_prompt_template_renders_commit_variant_with_title() {
-        assert_eq!(
-            review_prompt(
-                &ReviewTarget::Commit {
-                    sha: "deadbeef".to_string(),
-                    title: Some("Fix bug".to_string()),
-                },
-                &AbsolutePathBuf::current_dir().expect("cwd"),
-            )
-            .expect("commit prompt should render"),
-            "Review the code changes introduced by commit deadbeef (\"Fix bug\"). Provide prioritized, actionable findings."
-        );
-    }
-}
+#[path = "review_request_tests.rs"]
+mod review_request_tests;
