+++ codex-rs/prompts/templates/agents/hierarchical.md
