+++ codex-rs/prompts/templates/compact/prompt.md
