+++ codex-rs/prompts/templates/goals/budget_limit.md
