+++ codex-rs/prompts/templates/goals/continuation.md
