+++ codex-rs/prompts/templates/goals/objective_updated.md
