+++ codex-rs/prompts/templates/permissions/approval_policy/never.md
