+++ codex-rs/prompts/templates/permissions/approval_policy/on_request_rule_request_permission.md
