+++ codex-rs/prompts/templates/permissions/sandbox_mode/danger_full_access.md
