+++ codex-rs/prompts/templates/permissions/sandbox_mode/read_only.md
