+++ codex-rs/prompts/templates/permissions/sandbox_mode/workspace_write.md
