+++ codex-rs/prompts/templates/realtime/backend_prompt.md
