+++ codex-rs/prompts/templates/realtime/realtime_start.md
