+++ codex-rs/prompts/templates/review/exit_success.xml
